+++ service-worker.js
@@ -1,4 +1,4 @@
-const CACHE_NAME = "nk-pro-v99-4-17";
+const CACHE_NAME = "nk-pro-v99-4-17-ap14";
 const APP_SHELL = [
   "./",
   "./index.html",